1.clean dependency scope and refact the code to adjust that dependencies which has bean 'remove-changed' to 'provide' scope 2. refact ORM.Page 3.add Application,Session scope object to the MVC.request.model 4. etc.

+++ src/main/java/org/eweb4j/orm/Page.java
@@ -27,12 +27,12 @@ public interface Page<T> {
 	/**
 	 * Return the total row count for all pages.
 	 */
-    public int getTotalRowCount();
+    public long getTotalRowCount();
     
     /**
      * Return the total number of pages.
      */
-    public int getTotalPageCount();
+    public long getTotalPageCount();
     
     /**
      * Return the index position of this page.
